requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
EbookLib==0.18
feedparser==6.0.11
filelock==3.16.1
idna==3.10
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.4.1
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.2.1
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.6
tk==0.1.0
tldextract==5.1.3
tqdm==4.67.1
ttkbootstrap==1.10.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
